requirements.txt
Flask==2.3.2
pymongo==4.3.3
paramiko==3.4.0
matplotlib==3.7.1
apscheduler==3.10.1
python-dotenv==1.0.0
pytest==7.4.0